[level 1] Title: 두 정수 사이의 합, Time: 15.30 ms, Memory: 85.6 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -1,7 +1,7 @@
 class Solution {
     public long solution(int a, int b) {
         long result = 0;
-        for (long i = Math.min(a, b); i <= Math.max(a, b); i++) {
+        for (int i = Math.min(a, b); i <= Math.max(a, b); i++) {
             result += i;
         }
         return result;
